Make elasticsearch optional
