+++ Cargo.toml
@@ -22,8 +22,6 @@ crossbeam = "0.7.2"
 crossbeam-channel = "0.3.9"
 crossbeam-queue = "0.1.2"
 crossbeam-utils = "0.6.6"
-elastic = "~0.21.0-pre.5"
-elastic_derive = "~0.21.0-pre.5"
 fern = { git = "https://github.com/daboross/fern", rev = "1c8fd50", features = ["reopen-03"]}
 futures = "0.1.28"
 #notify = "5.0.0-pre.1"
@@ -31,11 +29,17 @@ notify = { git = "https://github.com/passcod/notify", rev = "e04ec0236b2" }  # f
 libc = "0.2.60"
 log = "^0.4"
 reopen = "^0.3.0"
-serde = "~1"
-serde_derive = "~1"
-serde_json = "~1"
 syslog = "4.0.1"
 signal-hook = "^0.1.10"
+
+elastic = { version = "~0.21.0-pre.5", optional = true }
+elastic_derive = { version = "~0.21.0-pre.5", optional = true }
+serde = { version = "~1", optional = true }
+serde_derive = { version = "~1", optional = true }
+serde_json = { version = "~1", optional = true }
+
+[features]
+elasticsearch-7 = ["elastic", "elastic_derive", "serde", "serde_derive", "serde_json"]
 
 [dev-dependencies]
 tempfile = "3.0.7"
@@ -48,3 +52,5 @@ sarchive = { path = "/usr/bin/sarchive" }
 
 [package.metadata.rpm.files]
 service = { path = "/usr/lib/systemd/system/sarchive.service" }
+
+maintenance = { status = "actively-developed" }
