+++ src/slurm.rs
@@ -19,8 +19,6 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
-use std::collections::HashMap;
-use std::fs;
 use std::path::{Path, PathBuf};
 use std::time::Instant;
 
@@ -41,28 +39,6 @@ impl SlurmJobEntry {
             jobid: id.to_string(),
             moment: Instant::now(),
         }
-    }
-
-    pub fn read_script(&self) -> String {
-        fs::read_to_string(self.path.join("script"))
-            .unwrap_or_else(|_| panic!("Could not read file {:?}/script", self.path))
-    }
-
-    pub fn read_env(&self) -> HashMap<String, String> {
-        let s = fs::read_to_string(self.path.join("environment"))
-            .unwrap_or_else(|_| panic!("Could not read file {:?}/environment", self.path));
-
-        s.split('\0')
-            .filter(|s| !s.is_empty())
-            .map(|s| {
-                let ps: Vec<_> = s.split('=').collect();
-                if ps.len() == 2 {
-                    (ps[0].to_owned(), ps[1].to_owned())
-                } else {
-                    (s.to_owned(), String::from(""))
-                }
-            })
-            .collect()
     }
 }
